arch(W3-T2): extract TurnEngine kernel behind Agent facade (behavior-preserving)
